requirements.txt
boto3==1.39.4
botocore==1.39.4
jmespath==1.0.1
pika==1.3.2
pillow==11.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.13.0
six==1.17.0
urllib3==2.5.0
